+++ lib/src/models.dart
@@ -1,4 +1,6 @@
+export 'data/ntp_inventory.dart' show curatedNtpInventory;
 export 'models/confidence_level.dart';
+export 'models/ntp_server_info.dart';
 export 'models/sync_metrics.dart';
 export 'models/trust_anchor.dart';
 export 'models/trust_anchor_contributor.dart';